Move `TextensionConsole.cs` into it's appropriate directory (Editor/Utilities)
